% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageFNN.R
\docType{package}
\name{FNN-package}
\alias{FNN}
\alias{FNN-package}
\title{FNN: Functional Neural Networks}
\description{
This package fits models of the form y(z, f(x) | theta) = z + f(x) + epsilon. In other words, this package will allow users to build
             models that have either functional or scalar responses paired with functional and scalar covariates. The main backbone of these models is
             neural networks. There are also a number of helper functions that make workflow easier for the user.
}
\details{
You can list off functions you wish to link to (maybe all of the functions in your package)
using the bracket syntax: \code{\link[=colSums]{colSums()}} for example.  \verb{Type like this} to get console font.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arxiv.org/abs/2006.09590}
}

}
\author{
\strong{Maintainer}: Barinder Thind \email{barinder.thi@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Jiguo Cao - aut Last
}

Other contributors:
\itemize{
  \item Richard Groenewald [contributor]
  \item Sidi Wu - ctb [contributor]
}

}
\keyword{internal}
