\name{print.knnReg}
\alias{print.knnRegCV}
\alias{print.knnReg}

\title{Print Method for KNN Regression}

\description{Print method for KNN regression.}

\usage{
\method{print}{knnReg}(x, ...)
\method{print}{knnRegCV}(x, ...)
}

\arguments{
  \item{x}{a \code{knnReg} or \code{knnRegCV} object.}
  \item{\dots}{Additonal \code{print} arguments.}
}

\author{Shengqiao Li. To report any bugs or suggestions please email: \email{shli@stat.wvu.edu.}}

\keyword{print}

