\name{FNN-package}
\alias{FNN-package}
\alias{FNN}
\docType{package}
\title{
Fast Nearest Neighbor Search Algorighms and Applications
}
\description{
 A collection of fast k-nearest neighbor search algorighms and applications
 including a  cover-tree, kd-tree and the nearest neighbor algorithm in package class. 
 Cover-tree templated code is taken from \url{http://hunch.net/~jl/projects/cover_tree/cover_tree.html}.
 The ANN C++ library  for kd-tree is taken from \url{http://www.cs.umd.edu/~mount/ANN/}. 
 In addtion, KNN classification and regression are implemented.
}
\details{
\tabular{ll}{
Package: \tab FNN\cr
Version: \tab 0.3-1\cr
Date: \tab 2010-3-16\cr
Depends: \tab R (>= 2.9.0)\cr
License: \tab GPL (>=2)\cr
Packaged: \tab 2010-03-17 01:24:09 UTC; Tiger\cr
Built: \tab R 2.10.1; i386-pc-mingw32; 2010-03-17 01:24:10 UTC; windows\cr
}

Index:
\preformatted{
KL.dist                 Kullback-Leibler Distance
get.knn                 Search Nearest Neighbors
knn                     k-Nearest Neighbour Classification
knn.cv                  k-Nearest Neighbour Classification
                        Cross-Validation
knn.reg                 k Nearest Neighbor Regression
}
}
\author{
Shengqiao Li <shli@stat.wvu.edu>

Maintainer: Shengqiao Li <shli@stat.wvu.edu>
}
\keyword{ package }
